default缺省语义实现，导入导出功能优化

+++ vrmCheck/src/main/java/com/nuaa/art/vrmcheck/model/table/AndOrConditionsInformation.java
@@ -27,7 +27,7 @@ public class AndOrConditionsInformation {
 
     public ArrayList<Integer> rowsForTrueScenarioSet; //永真行
     public ArrayList<Integer> rowsForDefaultScenarioSet; //默认行
-    public ArrayList<HashSet<Integer>> equivalentScenarioSet;// 每个场景对应的输出值号， V2版本为 每行对应的等价场景集
+    public ArrayList<HashSet<Long>> equivalentScenarioSet;// 每个场景对应的输出值号， V2版本为 每行对应的等价场景集
     public int[] variableRanges;
     public ScenarioCorpusCoder scenarioCorpusCoder;
